requirements.txt
Flask==2.3.0
torch==2.0.1
torchvision==0.15.0
Pillow==9.3.0
numpy==1.24.2
opencv-python==4.7.0
matplotlib==3.7.1
pytorch-grad-cam==1.6.0
werkzeug==2.3.0